Resolve items other than Quoted
