+++ compiler/noirc_frontend/src/hir_def/types.rs
@@ -2092,6 +2092,13 @@ impl Type {
             Type::Forall(_, typ) => typ.replace_named_generics_with_type_variables(),
         }
     }
+
+    pub fn slice_element_type(&self) -> Option<&Type> {
+        match self {
+            Type::Slice(element) => Some(element),
+            _ => None,
+        }
+    }
 }
 
 /// Wraps a given `expression` in `expression.as_slice()`
